requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
pdfplumber==0.11.0
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.2.1
pytest==8.3.3
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
